Blended with previous versions and made formating awesome.
Okay. So there were two previous versions of the beginning of
the story, (i.e one was when this whole ordeal first started over
IM, and the second was on google docs). Anyway, I liked them so
I have tried to incoperate them into this one. As a result, there
were minor modifications to your current version (I got rid of
some lines, and made it not-night etc). The two previous versions
can be found in the appendix in thestory.pdf.

I also made the formating look cool. I think it looks cool. =]